+++ WebApp/src/main/resources/db/migration/V1_10__RecipeProducts.sql
@@ -2,6 +2,6 @@ CREATE TABLE recipe_product(
   id            int PRIMARY KEY,
   recipe_id     int NOT NULL REFERENCES recipe(recipe_id),
   product_id    int NOT NULL REFERENCES product(id),
-  quantity      DECIMAL(8,3)
+  quantity      float
 );
 CREATE SEQUENCE recipe_product_seq;
